feat(vacaciones-pdf): redisenio completo formulario oficial
- Radio boxes con CSS puro (compatibles DomPDF)
- Motivo seleccionado resaltado en verde
- Fuente 12px mas legible
- Logo 108px
- Checkboxes en secciones de firma
- Fechas en formato d/m/Y
- Ocupa toda la hoja letter
- Periodos 2026 generados para todos los servidores

+++ sgth-backend/app/Services/Asistencia/PeriodoVacacionService.php
@@ -44,7 +44,7 @@ class PeriodoVacacionService
 
         if (!$fechaRef) return 0;
 
-        return Carbon::parse($fechaRef)
+        return (int) Carbon::parse($fechaRef)
             ->diffInYears(Carbon::create($anio, 12, 31));
     }
